requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.5.0
werkzeug==2.3.7
gunicorn==21.2.0
jinja2==3.1.2
requests==2.31.0 
httpx